Minor fix to UCC code for using set_parameters()
Essentially, need to get the coefficients (sign) of $\theta$ to use the set_parameters() method on the UCC circuit.
Otherwise, when trying to set the RZ parameter to be the same for each Pauli string in a fermionic excitation throughout, the CNOT ladders will cancel each other out, i.e. nothing happens!