Upload first draft of tutorial notebooks
First iteration of example/tutorial Jupyter notebooks uploaded; still
needs some editing to make the writing consistent throughout.
Also need to update the cells about calculating the expectation value of
an operator using the `expectation_from_samples` function.
Note: the `.gitattributes` file is supposed to remove the output of all
the Jupyter notebooks, need to check that it works.